+++ MoMoney/Converters/IconConverter.cs
@@ -16,19 +16,19 @@ class IconConverter : IMultiValueConverter
             {
                 var subcat = int.Parse(values[1].ToString());
 
-                return subcat == Constants.DEBIT_ID ? "grey_arrow_right.png" : "grey_arrow_left.png";
+                return subcat == Constants.DEBIT_ID ? "grey_arrow_right.svg" : "grey_arrow_left.svg";
             }
             else if (cat == Constants.INCOME_ID)
             {
-                return "green_arrow_up.png";
+                return "green_arrow_up.svg";
             }
             else if (cat >= Constants.EXPENSE_ID)
             {
-                return "red_arrow_down.png";
+                return "red_arrow_down.svg";
             }
         }
 
-        return "error.png";
+        return "error.svg";
     }
 
     public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)