+++ MoMoney/MauiProgram.cs
@@ -1,4 +1,5 @@
 ﻿using CommunityToolkit.Maui;
+using Microsoft.Extensions.Logging;
 using Syncfusion.Maui.Core.Hosting;
 using MoMoney.Data;
 using MoMoney.Views;
@@ -16,17 +17,9 @@ public static class MauiProgram
 	public static MauiApp CreateMauiApp()
 	{
 		var builder = MauiApp.CreateBuilder();
-		builder
-			.UseMauiApp<App>()
-            .ConfigureSyncfusionCore()
-            .UseMauiCommunityToolkit()
-            .ConfigureFonts(fonts =>
-			{
-				fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
-				fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
-				fonts.AddFont("Metropolis-Regular.otf", "Metropolis");
-				fonts.AddFont("Metropolis-Bold.otf", "MetropolisBold");
-			});
+		builder.UseMauiApp<App>()
+			   .ConfigureSyncfusionCore()
+			   .UseMauiCommunityToolkit();
 
         // db
         builder.Services.AddSingleton<MoMoneydb>();